layout/style/test/test_media_queries_dynamic_xbl.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=156716
-->
<head>
  <title>Test for Bug 156716</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="run()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=156716">Mozilla Bug 156716</a>
<iframe id="display" src="media_queries_dynamic_xbl_iframe.html"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 156716 **/

function run() {
  var iframe = document.getElementById("display");

  var subdoc = iframe.contentDocument;
  var subwin = iframe.contentWindow;
  var p = subdoc.getElementById("para");

  iframe.setAttribute("style", "height: 300px; width: 100px");
  is(subwin.getComputedStyle(p, "").color, "rgb(128, 0, 128)",
     "should be purple when portait");
  iframe.setAttribute("style", "height: 100px; width: 300px");
  is(subwin.getComputedStyle(p, "").color, "rgb(0, 0, 255)",
     "should be blue when landscape");
  SimpleTest.finish();
}
SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
